docs: docs for aa-ethers (wip)
