+++ site/packages/aa-ethers/utils/introduction.md
@@ -18,21 +18,20 @@ head:
 
 Notable util methods include:
 
-1.  [`convertWalletToAccountSigner`](/packages/aa-ethers/utils/convertWalletToAccountSigner) --
-2.  [`convertEthersSignerToAccountSigner`](/packages/aa-ethers/utils/convertEthersSignerToAccountSigner) --
+1.  [`convertWalletToAccountSigner`](/packages/aa-ethers/utils/convertWalletToAccountSigner) -- converts your ethers.js `Wallet` object into an `SmartAccountSigner` by deriving implementations of its methods.
+2.  [`convertEthersSignerToAccountSigner`](/packages/aa-ethers/utils/convertEthersSignerToAccountSigner) -- converts your ethers.js `Signer` object into an `SmartAccountSigner` by deriving implementations of its methods.
 
 ## Usage
 
 ::: code-group
 
 ```ts [example.ts]
-import { SupportedChains } from "@alchemy/aa-alchemy";
+// Wallet is a subclass of Signer, and so can be used with either convertor method
+const wallet = new Wallet(process.env.PRIVATE_KEY!);
+const accountSigner1 = convertWalletToAccountSigner(wallet);
 
-// eth mainnet
-const mainnet = SupportedChains.get(1);
-
-// bsc is unsupported, so the variable will be undefined
-const bsc = SupportedChains.get(56);
+// note that `signTypedData` is not supported by the Signer class, and so this util method cannot derive an implementation of said method for LocalAccountSigner
+const accountSigner2 = convertEthersSignerToAccountSigner(wallet);
 ```
 
 :::
